lesson_4

+++ src/main/java/com/igorkunicyn/market/controllers/OrderController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpSession;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Controller
@@ -25,7 +26,7 @@ public class OrderController {
     @GetMapping("")
     public String shopPage(Model model, HttpSession httpSession) {
         List<Product> productList = orderService.getListProducts(httpSession);
-        double price = orderService.totalPriceOrder(productList);
+        BigDecimal price = orderService.totalPriceOrder(productList);
         model.addAttribute("price",price);
         model.addAttribute("productList", productList);
         return "order-page";
